VM codegen changes:
  - Separate 32/64-bit code generation (TODO 64-bit).
  - Replace call to liblux for new instructions in the VM (this should speed things up a little).
  - The assembler and linker generate output for both the 32 and 64-bit VM.
  - The assembler now recognizes negative integer constants.

+++ src/decl.c
@@ -629,7 +629,7 @@ int is_typedef_name(char *id)
     return FALSE;
 }
 
-static long en_val = -1;
+static int en_val = -1;
 
 void reset_enum_val(void)
 {
